refactor and move to metricQueryUtils
