+++ frontend/src/scenes/experiments/metricQueryUtils.ts
@@ -21,6 +21,7 @@ import type {
     ExperimentFunnelMetric,
     ExperimentFunnelMetricStep,
     ExperimentMetric,
+    ExperimentMetricTypeProps,
     FunnelsFilter,
     FunnelsQuery,
     InsightVizNode,
@@ -241,7 +242,7 @@ const getFunnelSeries = (funnelMetric: ExperimentFunnelMetric): (EventsNode | Ac
 /**
  * Helper function to create a filter from a single source
  */
-const createFilterForSource = (source: ExperimentMetricSource): FilterType => {
+export const createFilterForSource = (source: ExperimentMetricSource): FilterType => {
     if (source.kind === NodeKind.EventsNode) {
         return {
             events: [createSourceNode(source)],
@@ -351,6 +352,117 @@ type ExperimentMetricSourceWithType =
     | (EventsNode & { type: 'events'; id: string; name: string })
     | (ActionsNode & { type: 'actions'; id: number; name: string })
     | (ExperimentDataWarehouseNode & { type: 'data_warehouse'; id: string; name: string })
+
+/**
+ * Converts filter data to a metric source (EventsNode, ActionsNode, or ExperimentDataWarehouseNode)
+ * Used for ratio metrics and mean metrics to support all source types
+ */
+export function filterToMetricSource(
+    actions: Record<string, any>[] | undefined,
+    events: Record<string, any>[] | undefined,
+    data_warehouse: Record<string, any>[] | undefined
+): ExperimentMetricSource | null {
+    if (events?.[0]) {
+        return {
+            kind: NodeKind.EventsNode,
+            event: events[0].id,
+            name: events[0].name,
+            math: events[0].math || ExperimentMetricMathType.TotalCount,
+            math_property: events[0].math_property,
+            math_hogql: events[0].math_hogql,
+            properties: events[0].properties,
+        }
+    }
+
+    if (actions?.[0]) {
+        return {
+            kind: NodeKind.ActionsNode,
+            id: actions[0].id,
+            name: actions[0].name,
+            math: actions[0].math || ExperimentMetricMathType.TotalCount,
+            math_property: actions[0].math_property,
+            math_hogql: actions[0].math_hogql,
+            properties: actions[0].properties,
+        }
+    }
+
+    if (data_warehouse?.[0]) {
+        return {
+            kind: NodeKind.ExperimentDataWarehouseNode,
+            name: data_warehouse[0].name,
+            table_name: data_warehouse[0].id,
+            timestamp_field: data_warehouse[0].timestamp_field,
+            events_join_key: data_warehouse[0].events_join_key,
+            data_warehouse_join_key: data_warehouse[0].data_warehouse_join_key,
+            math: data_warehouse[0].math || ExperimentMetricMathType.TotalCount,
+            math_property: data_warehouse[0].math_property,
+            math_hogql: data_warehouse[0].math_hogql,
+            properties: data_warehouse[0].properties,
+        }
+    }
+
+    return null
+}
+
+/**
+ * Converts filter format to metric configuration
+ * This is the reverse operation of getFilter
+ */
+export function filterToMetricConfig(
+    metricType: ExperimentMetricType,
+    actions: Record<string, any>[] | undefined,
+    events: Record<string, any>[] | undefined,
+    data_warehouse: Record<string, any>[] | undefined
+): ExperimentMetricTypeProps | undefined {
+    return match(metricType)
+        .with(ExperimentMetricType.FUNNEL, () => {
+            // Combine events and actions and sort by order
+            const eventSteps =
+                events?.map(
+                    (event) =>
+                        ({
+                            kind: NodeKind.EventsNode,
+                            event: event.id,
+                            properties: event.properties,
+                            order: event.order,
+                        }) as EventsNode & { order: number }
+                ) || []
+
+            const actionSteps =
+                actions?.map(
+                    (action) =>
+                        ({
+                            kind: NodeKind.ActionsNode,
+                            id: action.id,
+                            name: action.name,
+                            properties: action.properties,
+                            order: action.order,
+                        }) as ActionsNode & { order: number }
+                ) || []
+
+            const combinedSteps = [...eventSteps, ...actionSteps].sort((a, b) => a.order - b.order)
+
+            // Remove the temporary order field
+            const series = combinedSteps.map(({ order, ...step }) => step as ExperimentFunnelMetricStep)
+
+            return series.length > 0
+                ? {
+                      metric_type: ExperimentMetricType.FUNNEL as const,
+                      series,
+                  }
+                : undefined
+        })
+        .with(ExperimentMetricType.MEAN, () => {
+            const source = filterToMetricSource(actions, events, data_warehouse)
+            return source
+                ? {
+                      metric_type: ExperimentMetricType.MEAN as const,
+                      source,
+                  }
+                : undefined
+        })
+        .otherwise(() => undefined)
+}
 
 /**
  * this is a type adapter between metrics and filters.
